+++ MatrixSDK/MXSDKOptions.h
@@ -128,6 +128,13 @@ NS_ASSUME_NONNULL_BEGIN
 */
 @property (nonatomic, nullable) NSDictionary<NSString *, NSString*> *HTTPAdditionalHeaders;
 
+/**
+ Flag to automatically accept room invites.
+ 
+ @remark NO by default.
+ */
+@property (nonatomic, assign) BOOL autoAcceptRoomInvites;
+
 @end
 
 NS_ASSUME_NONNULL_END